+++ src/CodeWF.AvaloniaControls.Themes/CodeWFTheme.axaml.cs
@@ -1,0 +1,7 @@
+﻿using Avalonia.Styling;
+
+namespace CodeWF.AvaloniaControls.Themes;
+
+public class CodeWFTheme : Styles
+{
+}