+++ src/CodeWF.AvaloniaControls.Themes/Properties/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+using Avalonia.Metadata;
+
+[assembly: XmlnsPrefix("https://codewf.com", "codewf")]
+[assembly: XmlnsDefinition("https://codewf.com", "CodeWF.AvaloniaControls.Themes")]
+[assembly: XmlnsDefinition("https://codewf.com", "CodeWF.AvaloniaControls.Themes.Converters")]