+++ src/CodeWF.AvaloniaControls/AssemblyInfo.cs
@@ -2,6 +2,5 @@
 
 [assembly: XmlnsPrefix("https://codewf.com", "CodeWF.AvaloniaControls")]
 [assembly: XmlnsDefinition("https://codewf.com", "CodeWF.AvaloniaControls")]
-[assembly: XmlnsDefinition("https://codewf.com", "CodeWF.AvaloniaControls.Converters")]
 [assembly: XmlnsDefinition("https://codewf.com", "CodeWF.AvaloniaControls.Controls")]
 [assembly: XmlnsDefinition("https://codewf.com", "CodeWF.AvaloniaControls.Models")]