+++ src/CodeWF.AvaloniaControls/CodeWFTheme.axaml.cs
@@ -1,7 +1,0 @@
-﻿using Avalonia.Styling;
-
-namespace CodeWF.AvaloniaControls;
-
-public class CodeWFTheme : Styles
-{
-}