+++ src/CodeWF.Themes.StatusBadgeDemo/ViewModels/ViewModelBase.cs
@@ -1,8 +1,0 @@
-using ReactiveUI;
-
-namespace CodeWF.Themes.StatusBadgeDemo.ViewModels
-{
-    public class ViewModelBase : ReactiveObject
-    {
-    }
-}