+++ src/CodeWF.Themes/Index.axaml.cs
@@ -1,7 +1,0 @@
-﻿using Avalonia.Styling;
-
-namespace CodeWF.Themes;
-
-public class Theme : Styles
-{
-}