+++ src/CodeWF.Themes/Properties/AssemblyInfo.cs
@@ -1,5 +1,0 @@
-﻿using Avalonia.Metadata;
-
-[assembly:XmlnsPrefix("https://codewf.com", "c")]
-[assembly:XmlnsDefinition("https://codewf.com", "CodeWF.Themes")]
-[assembly: XmlnsDefinition("https://codewf.com", "CodeWF.Themes.Converters")]