+++ src/CodeWF/Properties/AssemblyInfo.cs
@@ -1,5 +1,0 @@
-using Avalonia.Metadata;
-
-[assembly:XmlnsPrefix("https://codewf.com", "c")]
-[assembly:XmlnsDefinition("https://codewf.com", "CodeWF")]
-[assembly:XmlnsDefinition("https://codewf.com", "CodeWF.Controls")]